Consolidate and hide validation gen. API
